fixed recent issues with the MultiKeyDictionary, multi-client should be possible again. report any bugs.

+++ trunk/Server2011/GWLP-R/ServerEngine/DataManagement/IWrapper.cs
@@ -6,7 +6,7 @@
                 ///   Explicit ToString override, because wrapper structs' ToString will hide the properties
                 /// </summary>
                 /// <returns></returns>
-                int Hash();
+                ulong Hash();
         }
 
         public interface IWrapper<out T> : IWrapper
